refactor: Move pairindextypes to sphinx.domains.python
